Move permissions helper to separate file
